+++ js/const.js
@@ -1,0 +1,42 @@
+const DESCRIPTIONS = [
+  'Вся красота мира в одной картинке',
+  'Моменты, которые запечатлены навсегда',
+  'Счастье в каждом кадре',
+  'Когда слова не нужны, достаточно фотографии',
+  'История, рассказанная через объектив',
+  'Остановить время в одном кадре',
+  'Фотография — это способ улыбнуться в будущем',
+  'Сегодня — самый лучший день',
+  'Я не доверяю словам. Я доверяю фотографиям',
+  'Фотографии — это свидетельство о том, что мы жили',
+  'Момент, когда небо и земля сливаются воедино',
+  'В объектив всегда видна правда — это как детектор лжи',
+  'Сделано объективом и любовью',
+  'Счастье никогда не выходит из моды',
+  'Лишь тот, кто странствует, открывает новые пути',
+  'Зарядитесь нашим теплом',
+  'Жизнь лучше, когда ты смеешься'
+];
+
+const MESSAGES = [
+  ' Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Михаил',
+  'Артём',
+  'Александр',
+  'Матвей',
+  'Максим',
+  'Елена',
+  'Наталья',
+  'Марина',
+  'Светлана'
+];
+
+export { DESCRIPTIONS, MESSAGES, NAMES };
