+++ js/util.js
@@ -1,0 +1,14 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+
+const counter = () => {
+  let n = 0;
+  return () => ++n;
+};
+
+export { getRandomInteger, counter };
